requirements.txt
Pillow
av
tkinterdnd2
opencv-python
pygame